pyproject.toml
[tool.poetry]
name = "testapp"
version = "0.1.0"
description = "Testing app"
authors = ["Dmitry"]

[tool.poetry.dependencies]
python = "^3.10"
pyleniumio = "^1.19.1"

[tool.poetry.dev-dependencies]
autopep8 = "^2.0.2"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
